refactor(47): renamed IsFinished to Status
